+++ plugins/guests/solaris11/cap/rsync.rb
@@ -20,7 +20,7 @@ module VagrantPlugins
           su_cmd = machine.config.solaris11.su_cmd
           machine.communicate.execute(
             "#{su_cmd} '#{opts[:guestpath]}' '(' ! -user #{opts[:owner]} -or ! -group #{opts[:group]} ')' -print0 | " +
-              "xargs -0 -r chown -v #{opts[:owner]}:#{opts[:group]}")
+              "xargs -0 -r chown #{opts[:owner]}:#{opts[:group]}")
         end
       end
     end